Add strict warning checks to tests
